+++ packages/collection/src/i18n.mjs
@@ -38,6 +38,7 @@ import { i18n as opal } from '@freesewing/opal'
 import { i18n as otis } from '@freesewing/otis'
 import { i18n as paco } from '@freesewing/paco'
 import { i18n as penelope } from '@freesewing/penelope'
+import { i18n as sabrina } from '@freesewing/sabrina'
 import { i18n as sandy } from '@freesewing/sandy'
 import { i18n as shelly } from '@freesewing/shelly'
 import { i18n as shin } from '@freesewing/shin'
@@ -101,6 +102,7 @@ export const i18n = {
   otis,
   paco,
   penelope,
+  sabrina,
   sandy,
   shelly,
   shin,
@@ -119,5 +121,5 @@ export const i18n = {
   wahid,
   walburga,
   waralee,
-  yuri
+  yuri,
 }