+++ packages/react/components/LineDrawing/index.mjs
@@ -29,6 +29,7 @@ import { Lumina, LuminaFront, LuminaBack } from './lumina.mjs'
 import { Lumira, LumiraFront, LumiraBack } from './lumira.mjs'
 import { Lunetius, LunetiusFront } from './lunetius.mjs'
 import { Noble, NobleFront, NobleBack } from './noble.mjs'
+import { Sabrina, SabrinaFront, SabrinaBack } from './sabrina.mjs'
 import { Simon, SimonFront, SimonBack } from './simon.mjs'
 import { Teagan, TeaganFront, TeaganBack } from './teagan.mjs'
 import { Tristan, TristanFront, TristanBack } from './tristan.mjs'
@@ -76,6 +77,7 @@ export const lineDrawingsFront = {
   lunetius: LunetiusFront,
   noble: NobleFront,
 
+  sabrina: SabrinaFront,
   simon: SimonFront,
 
   teagan: TeaganFront,
@@ -114,6 +116,7 @@ export const lineDrawingsBack = {
   lumina: LuminaBack,
   lumira: LumiraBack,
   noble: NobleBack,
+  sabrina: SabrinaBack,
   simon: SimonBack,
   teagan: TeaganBack,
   tristan: TristanBack,
@@ -160,6 +163,7 @@ export const lineDrawings = {
   lumira: Lumira,
   lunetius: Lunetius,
   noble: Noble,
+  sabrina: Sabrina,
   simon: Simon,
   teagan: Teagan,
   tristan: Tristan,