+++ packages/react/hooks/useDesign/index.mjs
@@ -37,6 +37,7 @@ import { Opal as opal } from '@freesewing/opal'
 import { Otis as otis } from '@freesewing/otis'
 import { Paco as paco } from '@freesewing/paco'
 import { Penelope as penelope } from '@freesewing/penelope'
+import { Sabrina as sabrina } from '@freesewing/sabrina'
 import { Sandy as sandy } from '@freesewing/sandy'
 import { Shelly as shelly } from '@freesewing/shelly'
 import { Shin as shin } from '@freesewing/shin'
@@ -97,6 +98,7 @@ export const designs = {
   otis,
   paco,
   penelope,
+  sabrina,
   sandy,
   shelly,
   shin,
@@ -115,7 +117,7 @@ export const designs = {
   wahid,
   walburga,
   waralee,
-  yuri
+  yuri,
 }
 
 export const useDesign = (design) => (designs[design] ? designs[design] : false)