+++ packages/react/components/Account/Bio.mjs
@@ -50,7 +50,7 @@ export const Bio = ({ welcome = false, Link = false }) => {
   const nextHref =
     welcomeSteps[account.control].length > 5
       ? '/welcome/' + welcomeSteps[account.control][6]
-      : '/docs/about/guide'
+      : '/docs/about/'
 
   return (
     <div className="tw:w-full">