+++ packages/react/components/Account/Compare.mjs
@@ -64,7 +64,7 @@ export const Compare = ({ welcome = false }) => {
   const nextHref =
     welcomeSteps[account?.control].length > 3
       ? '/welcome/' + welcomeSteps[account?.control][4]
-      : '/docs/about/guide'
+      : '/docs/about/'
 
   return (
     <div className="tw:w-full">