+++ CHANGELOG.md
@@ -3,11 +3,20 @@
 
 ## Unreleased
 
+
+## 4.1.0 (2025-06-29)
+
 ### octoplushy
 
 #### Fixed
 
  - Use full translation keys (#454)
+
+### sabrina
+
+#### Added
+
+ - Initial release of the Sabrina design
 
 ### core
 
@@ -30,7 +39,7 @@
  - Show snapped value for snapped percentage options with toAbs() set
 
 
-## 4.0.3 (2024-06-09)
+## 4.0.3 (2025-06-09)
 
 ### studio
 
@@ -39,7 +48,7 @@
  - Include signup / signin pages in studio
 
 
-## 4.0.2 (2024-06-09)
+## 4.0.2 (2025-06-09)
 
 ### studio
 
@@ -48,7 +57,7 @@
  - Add @freesewing/plugin-svgattr as a dependency to the studio
 
 
-## 4.0.1 (2024-06-09)
+## 4.0.1 (2025-06-09)
 
 ### bee
 
@@ -220,7 +229,7 @@
  - Use a dedicated logo image for dark mode
 
 
-## 4.0.0 (2024-04-01)
+## 4.0.0 (2025-04-01)
 
 ### bibi